moved power meter abc class
